fixup: Moved new callback code to new base/AsyncCallbacks.h
Also found the right place for EnableIfType, addressing branch XXX.
