+++ src/Downloader.h
@@ -9,6 +9,7 @@
 #ifndef SQUID_DOWNLOADER_H
 #define SQUID_DOWNLOADER_H
 
+#include "base/AsyncCallbacks.h"
 #include "base/AsyncJob.h"
 #include "defines.h"
 #include "http/forward.h"
