+++ src/adaptation/icap/Xaction.cc
@@ -13,6 +13,7 @@
 #include "adaptation/icap/Config.h"
 #include "adaptation/icap/Launcher.h"
 #include "adaptation/icap/Xaction.h"
+#include "base/AsyncCallbacks.h"
 #include "base/JobWait.h"
 #include "base/TextException.h"
 #include "comm.h"
