+++ src/base/forward.h
@@ -25,6 +25,7 @@ template <typename Value> class Optional;
 template<class Cbc> class CbcPointer;
 template<class RefCountableKid> class RefCount;
 template<class Job> class JobWait;
+template<class Answer> class AsyncCallback;
 
 typedef CbcPointer<AsyncJob> AsyncJobPointer;
 typedef RefCount<CodeContext> CodeContextPointer;
