+++ src/htcp.cc
@@ -12,7 +12,7 @@
 #include "AccessLogEntry.h"
 #include "acl/Acl.h"
 #include "acl/FilledChecklist.h"
-#include "base/AsyncCbdataCalls.h"
+#include "base/AsyncCallbacks.h"
 #include "CachePeer.h"
 #include "comm.h"
 #include "comm/Connection.h"
